computes props in backend lib, handles period filter
