+++ src/components/shared/orgunit_picker/orgUnitTreeBackend.js
@@ -1,6 +1,12 @@
 import { getInstance } from "d2/lib/d2";
 import PluginRegistry from "../../core/PluginRegistry";
 
+let currentPeriod = undefined;
+
+export const setPeriod = (argPeriod) => {
+  currentPeriod = argPeriod;
+};
+
 const loadedOrgUnitsById = {};
 let contractsByOrgUnitId = {};
 const defaultOrgUnitFields =
@@ -27,7 +33,7 @@ export const getRootData = async (id, type = "source") => {
 const withHasChildren = (organisationUnits) => {
   for (let ou of organisationUnits) {
     ou.has_children = ou.children && ou.children.length > 0;
-    ou.activeContracts = contractsByOrgUnitId[ou.id];
+    ou.activeContracts = (contractsByOrgUnitId[ou.id] || []).filter((c) => c.matchPeriod(currentPeriod));
 
     loadedOrgUnitsById[ou.id] = ou;
   }
@@ -60,4 +66,48 @@ export const request = async (value, count, source, version) => {
   });
 
   return withHasChildren(resp.organisationUnits);
+};
+
+export const label = (data) => {
+  return data.name;
+};
+
+export const search = (input1, input2, type) => {
+  console.log(input1, input2, type);
+  return [];
+};
+
+export const makeDropDownText = (orgUnit) => {
+  return (
+    <div>
+      <span display="block">
+        {orgUnit.name} <code style={{ color: "lightgrey" }}>{orgUnit.id}</code>
+      </span>
+      <pre style={{ fontSize: "8px" }}>
+        {orgUnit.ancestors
+          .slice(1)
+          .map((o) => o.name)
+          .join(" > ")}
+      </pre>
+      <hr />
+    </div>
+  );
+};
+
+const parseNodeIds = (orgUnit) => {
+  const parsed = orgUnit.ancestors.map((a) => [a.id, a]).concat([[orgUnit.id, orgUnit]]);
+  return new Map(parsed);
+};
+
+
+export const treeProps = {
+  getRootData,
+  label,
+  getChildrenData,
+  search,
+  request,
+  makeDropDownText,
+  parseNodeIds: parseNodeIds,
+  toggleOnLabelClick: false,
+  isSelectable: () => true,
 };
